fix: address CodeRabbit review comments
- Split test file into per-module files (test_models.py + test_migration_add_pr_creators_index.py)
- Remove brittle index count test
- Unify duplicate upgrade/downgrade test helpers
- Add DROP INDEX before CREATE to handle invalid indexes from failed CONCURRENTLY builds
- Add BEGIN after CONCURRENTLY DDL to restore Alembic transaction context
- Use Literal type for migration action parameter
- Use more robust alembic.op.get_bind patching
